ci: fix golangci-lint config
The config was enforced to use Go 1.20 while the go.mod file is using Go 1.22.

golangci-lint detects the version it has to use from the go.mod file.
so it's not necessary to enforce it in the config.

Also, there is typecheck linter. typecheck is not a linter but an analyzer.